Unexport the write lock; remote does not need it
WriteLock was the only public symbol this branch added, and a Windows staging
primitive is not something a caller of an astronomy library has any use for. It
is now writeLock, unexported, and Save is the door that holds it. The code stays
in remote/file rather than moving to internal: it belongs beside the write it
guards.

That was possible because remote's two calls turned out to be redundant, which
is a conclusion rather than a convenience. writeStaged and the promote copy are
reachable from exactly one place, and it runs holding acquireLock - exclusive
across processes through IfNotExist and, since #245, within one too. Two
different cache keys stage under two different names now that the bucket URL
carries no_tmp_dir=1, and two buckets are two directories. So nothing in that
path had a second writer to exclude. Both call sites are gone and the promote
helper with them, since it existed only to scope the lock.

What is left needing the lock is file.Save, which anyone can call concurrently
for anything - a sidecar, a decoded payload, three parallel tests seeding one
kernel name, which is what filed #241.

The exclusivity test moves in-package, since it now calls an unexported
function. It is worth keeping separate from the two concurrency tests beside
it: those exercise the lock through a write and can pass for the wrong reason.

Verified: no exported symbol is added to remote/file at all. The spk suite
stays at 0 failures in 8 runs, remote passes under -race, and removing the lock
from Save still fails TestConcurrentSaveOfOneKey.

One thing worth recording: docsguard's TestCitedSymbolsExist caught the stale
`file.WriteLock` citation left behind in this PR's own changelog fragment. That
is the check #268 argues for, catching its author.

Closes #241.
